+++ internal/gen/controller/api/services/auth_method_service.pb.go
@@ -8,8 +8,8 @@ package services
 
 import (
 	_ "github.com/grpc-ecosystem/grpc-gateway/v2/protoc-gen-openapiv2/options"
-	authmethods "github.com/hashicorp/boundary/internal/gen/controller/api/resources/authmethods"
-	authtokens "github.com/hashicorp/boundary/internal/gen/controller/api/resources/authtokens"
+	authmethods "github.com/hashicorp/boundary/sdk/pbs/controller/api/resources/authmethods"
+	authtokens "github.com/hashicorp/boundary/sdk/pbs/controller/api/resources/authtokens"
 	_ "google.golang.org/genproto/googleapis/api/annotations"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"